Move `JavaHashPasswordHasher` to `org.simple.clinic.security.pin`
